Fix tests in JS.
